backend/requirements.txt
# FastAPI Backend Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
authlib==1.3.0
python-dotenv==1.0.0

# AWS
boto3==1.34.34
botocore==1.34.34

# Web scraping (existing)
requests==2.31.0
beautifulsoup4==4.12.3
selenium==4.16.0
webdriver-manager==4.0.1
camoufox==0.4.4
apify-client==1.7.1

# AI
openai>=1.50.0

# LangChain for multi-agent system
langchain==0.1.0
langchain-openai==0.0.2
langchain-core==0.1.10
langgraph==0.0.20

# PDF generation
reportlab==4.0.9
pypdf2==3.0.1

# DOCX generation (NEW)
python-docx==1.1.0

# OAuth & Web Server
flask==3.0.0

# Utilities
colorama==0.4.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Redis (optional, for caching)
redis==5.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
